lkl: env: posix: implement user dump_stack() callback
You'll need -rdynamic added to the linker in the kernel to be able to
see kernel symbols too (though static symbol names won't be exported).

Note: backtrace() and co. are GNU extensions, so this violates the
posix.c claim to be fully POSIX compatible.

Support for these functions should be detected at compile time to
retain at least compile-time compatibility with POSIX.

Signed-off-by: Lucian Adrian Grijincu <lucian.grijincu@cs.pub.ro>