lkl: kernel: call do_exit() on threads that exit
Without this we get a deadlock in this situation:
1) mount an journaled ext3 fs. This will start the kjournald()
   kthread.
2) unmount the ext3 fs. This will wait until the kjournald() kthread
   terminates.

The kjournald() thread does no call do_exit() on it's exit
path. Because of this, the kernel function executed by this kthread
will return and the current thread will terminate WITHOUT passing the
control to another thread (_switch_to()).

The kernel is now dead-locked: the only thread that was running has
left the building, and all other threads are waiting on their
respective sched-semaphores.

Calling do_exit() on this thread, will lead to a call to schedule(),
with the ->dead bit set. This will properly exit the current thread,
and will signal the sched-sem of another thread, preventing the kernel
from getting stuck.

This is not speciffic to kjournald. Any kthread started the same way,
and which does not call do_exit() could have dead-locked the kernel.

Signed-off-by: Lucian Adrian Grijincu <lucian.grijincu@cs.pub.ro>