lkl: env: ntk: make the timer a SynchronizationTimer
Before the timer was an NotificationTimer. That kind of timer was
signaled once when the timer fired and remained fired until someone
explicitly unsignaled it. The nkt lkl-env never unsignaled the
timer.

This means that the while() loop that waited on the timer to be
signaled and triggered an timer IRQ never really waited and generated
timer IRQs very rapidly.

There causees two problems:
a) the CPU wasted too many cycles generating timer IRQs
b) the IRQ semaphore was released too many times.

In some situations (for example when an active kernel thread was
waiting or doing some slow work) the IRQ semaphore could be released
too many times, more than it advertised when created, leading to a BSOD.

By changing the timer's type to SynchronizationTimer, the timer gets
signaled once when the timer fires and gets unsignaled immediatly
after the first thread waiting on it is scheduled.

This means that the IRQ generating code really does what is thought it
was doing until now: it *waits* until the timer gets signaled, which
only happens once in a while, not always as with NotificationTimer.

Signed-off-by: Lucian Adrian Grijincu <lucian.grijincu@cs.pub.ro>